+++ src/lang/prog/embed/mod.rs
@@ -1,0 +1,17 @@
+// devela::lang::prog::embed
+//
+#![doc = crate::_DOC_LANG_PROG_EMBED!()] // public
+#![doc = crate::_doc!(modules: crate::lang::prog; embed)]
+#![doc = crate::_doc!(flat:"lang")]
+#![doc = crate::_doc!(hr)]
+//
+
+// mod script;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     script::_all::*,
+        // };
+    }
+}