+++ src/lang/prog/kernel/mod.rs
@@ -1,0 +1,23 @@
+// devela::lang::prog::kernel
+//
+#![doc = crate::_DOC_LANG_PROG_KERNEL!()] // public
+#![doc = crate::_doc!(modules: crate::lang::prog; kernel)]
+#![doc = crate::_doc!(flat:"lang")]
+#![doc = crate::_doc!(hr)]
+//
+
+// mod array; // APL/J/K/BQN/Uiua-like array languages
+// mod concat; // Forth/False/Joy/Factor-like concatenative machines
+// mod rule; // AWK/Make/Prolog-ish rule or pattern-action systems
+// mod symb; // Lisp/Scheme/Racket-like symbolic tree forms
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     array::_all::*,
+        //     concat::_all::*,
+        //     rule::_all::*,
+        //     symb::_all::*,
+        // };
+    }
+}