+++ src/lang/prog/phrase/mod.rs
@@ -1,0 +1,17 @@
+// devela::lang::prog::phrase
+//
+#![doc = crate::_DOC_LANG_PROG_PHRASE!()] // public
+#![doc = crate::_doc!(modules: crate::lang::prog; phrase)]
+#![doc = crate::_doc!(flat:"lang")]
+#![doc = crate::_doc!(hr)]
+//
+
+// mod kind;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     kind::_all::*,
+        // };
+    }
+}