+++ src/lang/prog/script/mod.rs
@@ -1,0 +1,31 @@
+// devela::lang::prog::script
+//
+#![doc = crate::_DOC_LANG_PROG_SCRIPT!()] // public
+#![doc = crate::_doc!(modules: crate::lang::prog; script)]
+#![doc = crate::_doc!(flat:"lang")]
+#![doc = crate::_doc!(hr)]
+//
+
+// pub mod apl;
+// pub mod awk;
+// pub mod cmd;
+// pub mod false;
+// pub mod forth;
+// pub mod j;
+// pub mod php;
+// pub mod scheme;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super:: {
+        //     apl::_all::*,
+        //     awk::_all::*,
+        //     cmd::_all::*,
+        //     false::_all::*,
+        //     forth::_all::*,
+        //     j::_all::*,
+        //     php::_all::*,
+        //     scheme::_all::*,
+        // };
+    }
+}